requirements.txt
google-generativeai>=0.3.0
python-docx>=1.1.0
markdown>=3.5.0
pypdf>=3.17.0
reportlab>=4.0.7
pyyaml>=6.0.1
python-dotenv>=1.0.0
